# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A custom PaaS platform for automated application deployment.\nDeploy applications from Git repositories to Kubernetes with zero-configuration.",
        "title": "App Deployer API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/alvesdmateus/app-deployer"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/deployments/{deployment_id}/infrastructure": {
            "get": {
                "description": "Returns the infrastructure details for a specific deployment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "infrastructure"
                ],
                "summary": "Get infrastructure details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment ID",
                        "name": "deployment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.InfrastructureResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.InfrastructureResponse": {
            "type": "object",
            "properties": {
                "cluster_name": {
                    "type": "string"
                },
                "config": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deployment_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}